feat!: sdk regeneration 2026-04-27 (#42)
* chore: initialize SDK regeneration branch

* chore: unfreeze files pending regen

* chore(repo): ignore agent docs and Claude skills in fern regen

* chore(regen): automate ClientOptions patch reapply

* SDK regeneration

* chore: re-apply manual patches after regen

  - Restore ClientOptions.java SDK header constants and release-please version markers (overwritten by Fern with com.deepgram.fern:api-sdk and version 0.2.2).
  - Restore context7.json and add it to .fernignore so future regens don't delete it.
  - Rewrite the Agent WebSocket README snippet for the upstream agent settings restructure: ThinkSettingsV1 and ThinkSettingsV1Provider replace the AgentV1SettingsAgentThinkOneItem wrappers, and OpenAiThinkProvider now requires .builder().model(...) instead of accepting a free-form Map.

* fix(examples): align Java examples with regenerated SDK types

* chore(regen): align Java patch workflow with backup-restore strategy

* chore(repo): dedupe agent entries in .fernignore

---------

+++ src/main/java/com/deepgram/resources/listen/v1/types/ListenV1Metadata.java
@@ -29,7 +29,7 @@ public final class ListenV1Metadata {
 
     private final double duration;
 
-    private final double channels;
+    private final int channels;
 
     private final Map<String, Object> additionalProperties;
 
@@ -39,7 +39,7 @@ public final class ListenV1Metadata {
             String sha256,
             String created,
             double duration,
-            double channels,
+            int channels,
             Map<String, Object> additionalProperties) {
         this.transactionKey = transactionKey;
         this.requestId = requestId;
@@ -102,7 +102,7 @@ public final class ListenV1Metadata {
      * @return The channels
      */
     @JsonProperty("channels")
-    public double getChannels() {
+    public int getChannels() {
         return channels;
     }
 
@@ -182,7 +182,7 @@ public final class ListenV1Metadata {
         /**
          * <p>The channels</p>
          */
-        _FinalStage channels(double channels);
+        _FinalStage channels(int channels);
     }
 
     public interface _FinalStage {
@@ -212,7 +212,7 @@ public final class ListenV1Metadata {
 
         private double duration;
 
-        private double channels;
+        private int channels;
 
         @JsonAnySetter
         private Map<String, Object> additionalProperties = new HashMap<>();
@@ -297,7 +297,7 @@ public final class ListenV1Metadata {
          */
         @java.lang.Override
         @JsonSetter("channels")
-        public _FinalStage channels(double channels) {
+        public _FinalStage channels(int channels) {
             this.channels = channels;
             return this;
         }
